Unite all async resamplers
