+++ examples/fastfixedin_ramp64.rs
@@ -1,5 +1,5 @@
 extern crate rubato;
-use rubato::{Fast, Fixed, PolynomialDegree, Resampler};
+use rubato::{Async, Fixed, PolynomialDegree, Resampler};
 use std::convert::TryInto;
 use std::env;
 use std::fs::File;
@@ -114,7 +114,7 @@ fn main() {
 
     let chunksize = 1024;
     let target_ratio = final_ratio / 100.0;
-    let mut resampler = Fast::<f64>::new(
+    let mut resampler = Async::<f64>::new_poly(
         f_ratio,
         target_ratio,
         PolynomialDegree::Cubic,
