+++ src/asynchro_fast.rs
@@ -1,10 +1,4 @@
-use std::fmt;
-
-use crate::error::{ResampleError, ResampleResult, ResamplerConstructionError};
-use crate::{update_mask_from_buffers, validate_buffers, Fixed, Resampler, Sample};
-
-const POLYNOMIAL_LEN_U: usize = 8;
-const POLYNOMIAL_LEN_I: isize = 8;
+use crate::Sample;
 
 macro_rules! t {
     // Shorter form of T::coerce(value)
@@ -29,78 +23,21 @@ pub enum PolynomialDegree {
     Nearest,
 }
 
-/// An asynchronous resampler that either takes a fixed number of input frames,
-/// or returns a fixed number of audio frames.
-///
-/// The `fixed` argument determines if input of output should be fixed size.
-/// When the input size is fixed, the output size varies from call to call,
-/// and when output size is fixed, the input size varies.
-///
-/// The number of frames on the fixed side is determined by the chunk size argument to the constructor.
-/// This value can be changed by the `set_chunk_size()` method,
-/// to let the resampler process smaller chunks of audio data.
-/// Note that the chunk size cannot exceed the value given at creation time.
-///
-/// When the input size is fixed, the maximum value can be retrieved using the `input_size_max()` method,
-/// and `input_frames_next()` gives the current value.
-/// When the output size is fixed, the corresponding values are instead provided by the `output_size_max()`
-/// and `output_size_next()` methods.
-///
-/// The resampling is done by interpolating between the input samples.
-/// The polynomial degree can be selected, see [PolynomialDegree] for the available options.
-///
-/// Note that no anti-aliasing filter is used.
-/// This makes it run considerably faster than the corresponding Sinc resampler, which performs anti-aliasing filtering.
-/// The price is that the resampling creates some artefacts in the output, mainly at higher frequencies.
-/// Use a Sinc resampler if this can not be tolerated.
-///
-/// The resampling ratio can be freely adjusted within the range specified to the constructor.
-/// Higher maximum ratios require more memory to be allocated by
-/// [input_buffer_allocate](Resampler::input_buffer_allocate),
-/// [output_buffer_allocate](Resampler::output_buffer_allocate), and an internal buffer.
-pub struct Fast<T> {
-    nbr_channels: usize,
-    chunk_size: usize,
-    max_chunk_size: usize,
-    needed_input_size: usize,
-    needed_output_size: usize,
-    last_index: f64,
-    current_buffer_fill: usize,
-    resample_ratio: f64,
-    resample_ratio_original: f64,
-    target_ratio: f64,
-    max_relative_ratio: f64,
-    buffer: Vec<Vec<T>>,
-    interpolation: PolynomialDegree,
-    channel_mask: Vec<bool>,
-    fixed: Fixed,
-}
-
-impl<T> fmt::Debug for Fast<T> {
-    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt.debug_struct("Fast")
-            .field("nbr_channels", &self.nbr_channels)
-            .field("chunk_size,", &self.chunk_size)
-            .field("max_chunk_size,", &self.max_chunk_size)
-            .field("needed_input_size,", &self.needed_input_size)
-            .field("needed_output_size,", &self.needed_output_size)
-            .field("last_index", &self.last_index)
-            .field("current_buffer_fill", &self.current_buffer_fill)
-            .field("resample_ratio", &self.resample_ratio)
-            .field("resample_ratio_original", &self.resample_ratio_original)
-            .field("target_ratio", &self.target_ratio)
-            .field("max_relative_ratio", &self.max_relative_ratio)
-            .field("buffer[0].len()", &self.buffer[0].len())
-            .field("interpolation", &self.interpolation)
-            .field("channel_mask", &self.channel_mask)
-            .field("fixed", &self.fixed)
-            .finish()
+impl PolynomialDegree {
+    pub fn len(&self) -> usize {
+        match self {
+            PolynomialDegree::Nearest => 1,
+            PolynomialDegree::Linear => 2,
+            PolynomialDegree::Cubic => 4,
+            PolynomialDegree::Quintic => 6,
+            PolynomialDegree::Septic => 8,
+        }
     }
 }
 
 /// Perform septic polynomial interpolation to get value at x.
 /// Input points are assumed to be at x = -3, -2, -1, 0, 1, 2, 3, 4.
-fn interp_septic<T>(x: T, yvals: &[T]) -> T
+pub fn interp_septic<T>(x: T, yvals: &[T]) -> T
 where
     T: Sample,
 {
@@ -149,7 +86,7 @@ where
 
 /// Perform quintic polynomial interpolation to get value at x.
 /// Input points are assumed to be at x = -2, -1, 0, 1, 2, 3.
-fn interp_quintic<T>(x: T, yvals: &[T]) -> T
+pub fn interp_quintic<T>(x: T, yvals: &[T]) -> T
 where
     T: Sample,
 {
@@ -175,7 +112,7 @@ where
 
 /// Perform cubic polynomial interpolation to get value at x.
 /// Input points are assumed to be at x = -1, 0, 1, 2.
-fn interp_cubic<T>(x: T, yvals: &[T]) -> T
+pub fn interp_cubic<T>(x: T, yvals: &[T]) -> T
 where
     T: Sample,
 {
@@ -189,911 +126,9 @@ where
 }
 
 /// Linear interpolation between two points at x=0 and x=1.
-fn interp_lin<T>(x: T, yvals: &[T]) -> T
+pub fn interp_lin<T>(x: T, yvals: &[T]) -> T
 where
     T: Sample,
 {
     yvals[0] + x * (yvals[1] - yvals[0])
-}
-
-fn validate_ratios(
-    resample_ratio: f64,
-    max_resample_ratio_relative: f64,
-) -> Result<(), ResamplerConstructionError> {
-    if resample_ratio <= 0.0 {
-        return Err(ResamplerConstructionError::InvalidRatio(resample_ratio));
-    }
-    if max_resample_ratio_relative < 1.0 {
-        return Err(ResamplerConstructionError::InvalidRelativeRatio(
-            max_resample_ratio_relative,
-        ));
-    }
-    Ok(())
-}
-
-impl<T> Fast<T>
-where
-    T: Sample,
-{
-    /// Create a new Fast.
-    ///
-    /// Parameters are:
-    /// - `resample_ratio`: Starting ratio between output and input sample rates, must be > 0.
-    /// - `max_resample_ratio_relative`: Maximum ratio that can be set with [Resampler::set_resample_ratio] relative to `resample_ratio`, must be >= 1.0. The minimum relative ratio is the reciprocal of the maximum. For example, with `max_resample_ratio_relative` of 10.0, the ratio can be set between `resample_ratio * 10.0` and `resample_ratio / 10.0`.
-    /// - `interpolation_type`: Degree of polynomial used for interpolation, see [PolynomialDegree].
-    /// - `chunk_size`: Size of input data in frames.
-    /// - `nbr_channels`: Number of channels in input/output.
-    /// - `fixed`: Deciding whether input or output size is fixed.
-    pub fn new(
-        resample_ratio: f64,
-        max_resample_ratio_relative: f64,
-        interpolation_type: PolynomialDegree,
-        chunk_size: usize,
-        nbr_channels: usize,
-        fixed: Fixed,
-    ) -> Result<Self, ResamplerConstructionError> {
-        debug!(
-            "Create new Fast with fixed {:?}, ratio: {}, chunk_size: {}, channels: {}",
-            fixed, resample_ratio, chunk_size, nbr_channels,
-        );
-
-        validate_ratios(resample_ratio, max_resample_ratio_relative)?;
-
-        let channel_mask = vec![true; nbr_channels];
-
-        let last_index = -(POLYNOMIAL_LEN_I / 2) as f64;
-        let needed_input_size = Self::calculate_input_size(
-            chunk_size,
-            resample_ratio,
-            resample_ratio,
-            last_index,
-            &fixed,
-        );
-        let needed_output_size = Self::calculate_output_size(
-            chunk_size,
-            resample_ratio,
-            resample_ratio,
-            last_index,
-            &fixed,
-        );
-
-        let buffer_len = Fast::<T>::calculate_max_input_size(
-            chunk_size,
-            resample_ratio,
-            max_resample_ratio_relative,
-            &fixed,
-        ) + 2 * POLYNOMIAL_LEN_U;
-        let buffer = vec![vec![T::zero(); buffer_len]; nbr_channels];
-
-        Ok(Fast {
-            nbr_channels,
-            chunk_size,
-            max_chunk_size: chunk_size,
-            needed_input_size,
-            needed_output_size,
-            current_buffer_fill: needed_input_size,
-            last_index,
-            resample_ratio,
-            resample_ratio_original: resample_ratio,
-            target_ratio: resample_ratio,
-            max_relative_ratio: max_resample_ratio_relative,
-            buffer,
-            interpolation: interpolation_type,
-            channel_mask,
-            fixed,
-        })
-    }
-
-    fn calculate_input_size(
-        chunk_size: usize,
-        resample_ratio: f64,
-        target_ratio: f64,
-        last_index: f64,
-        fixed: &Fixed,
-    ) -> usize {
-        match fixed {
-            Fixed::Input => chunk_size,
-            Fixed::Output => (last_index
-                + chunk_size as f64 / (0.5 * resample_ratio + 0.5 * target_ratio)
-                + POLYNOMIAL_LEN_U as f64)
-                .ceil() as usize,
-        }
-    }
-
-    fn calculate_output_size(
-        chunk_size: usize,
-        resample_ratio: f64,
-        target_ratio: f64,
-        last_index: f64,
-        fixed: &Fixed,
-    ) -> usize {
-        match fixed {
-            Fixed::Output => chunk_size,
-            Fixed::Input => ((chunk_size as f64 - (POLYNOMIAL_LEN_U + 1) as f64 - last_index)
-                * (0.5 * resample_ratio + 0.5 * target_ratio))
-                .floor() as usize,
-        }
-    }
-
-    fn calculate_max_input_size(
-        chunk_size: usize,
-        resample_ratio_original: f64,
-        max_relative_ratio: f64,
-        fixed: &Fixed,
-    ) -> usize {
-        match fixed {
-            Fixed::Input => chunk_size,
-            Fixed::Output => {
-                (chunk_size as f64 / resample_ratio_original * max_relative_ratio).ceil() as usize
-                    + 2
-                    + POLYNOMIAL_LEN_U / 2
-            }
-        }
-    }
-
-    fn calculate_max_output_size(
-        chunk_size: usize,
-        resample_ratio_original: f64,
-        max_relative_ratio: f64,
-        fixed: &Fixed,
-    ) -> usize {
-        match fixed {
-            Fixed::Output => chunk_size,
-            Fixed::Input => {
-                (chunk_size as f64 * resample_ratio_original * max_relative_ratio + 10.0) as usize
-            }
-        }
-    }
-
-    fn update_lengths(&mut self) {
-        self.needed_input_size = Fast::<T>::calculate_input_size(
-            self.chunk_size,
-            self.resample_ratio,
-            self.target_ratio,
-            self.last_index,
-            &self.fixed,
-        );
-        self.needed_output_size = Fast::<T>::calculate_output_size(
-            self.chunk_size,
-            self.resample_ratio,
-            self.target_ratio,
-            self.last_index,
-            &self.fixed,
-        );
-        trace!(
-            "Updated lengths, input: {}, output: {}",
-            self.needed_input_size,
-            self.needed_output_size
-        );
-    }
-}
-
-impl<T> Resampler<T> for Fast<T>
-where
-    T: Sample,
-{
-    fn process_into_buffer<Vin: AsRef<[T]>, Vout: AsMut<[T]>>(
-        &mut self,
-        wave_in: &[Vin],
-        wave_out: &mut [Vout],
-        active_channels_mask: Option<&[bool]>,
-    ) -> ResampleResult<(usize, usize)> {
-        if let Some(mask) = active_channels_mask {
-            self.channel_mask.copy_from_slice(mask);
-        } else {
-            update_mask_from_buffers(&mut self.channel_mask);
-        };
-
-        trace!("Start processing, {:?}", self);
-
-        validate_buffers(
-            wave_in,
-            wave_out,
-            &self.channel_mask,
-            self.nbr_channels,
-            self.needed_input_size,
-            self.needed_output_size,
-        )?;
-
-        let interpolator_len = POLYNOMIAL_LEN_U;
-
-        let mut t_ratio = 1.0 / self.resample_ratio;
-        let t_ratio_end = 1.0 / self.target_ratio;
-
-        let t_ratio_increment = (t_ratio_end - t_ratio) / self.needed_output_size as f64;
-
-        // Update buffer with new data.
-        for buf in self.buffer.iter_mut() {
-            buf.copy_within(
-                self.current_buffer_fill..self.current_buffer_fill + 2 * interpolator_len,
-                0,
-            );
-        }
-
-        for (chan, wave_in) in wave_in
-            .iter()
-            .enumerate()
-            .filter(|(chan, _)| self.channel_mask[*chan])
-        {
-            self.buffer[chan][2 * interpolator_len..2 * interpolator_len + self.needed_input_size]
-                .copy_from_slice(&wave_in.as_ref()[..self.needed_input_size]);
-        }
-
-        self.current_buffer_fill = self.needed_input_size;
-
-        let mut idx = self.last_index;
-
-        match self.interpolation {
-            PolynomialDegree::Septic => {
-                for n in 0..self.needed_output_size {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 3;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in self.channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = self.buffer.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 8) as usize,
-                                );
-                                *wave_out
-                                    .get_unchecked_mut(chan)
-                                    .as_mut()
-                                    .get_unchecked_mut(n) = interp_septic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Quintic => {
-                for n in 0..self.needed_output_size {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 2;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in self.channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = self.buffer.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 6) as usize,
-                                );
-                                *wave_out
-                                    .get_unchecked_mut(chan)
-                                    .as_mut()
-                                    .get_unchecked_mut(n) = interp_quintic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Cubic => {
-                for n in 0..self.needed_output_size {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize - 1;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in self.channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = self.buffer.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 4) as usize,
-                                );
-                                *wave_out
-                                    .get_unchecked_mut(chan)
-                                    .as_mut()
-                                    .get_unchecked_mut(n) = interp_cubic(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Linear => {
-                for n in 0..self.needed_output_size {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let idx_floor = idx.floor();
-                    let start_idx = idx_floor as isize;
-                    let frac = idx - idx_floor;
-                    let frac_offset = t!(frac);
-                    for (chan, active) in self.channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let buf = self.buffer.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize
-                                        ..(start_idx + 2 * interpolator_len as isize + 2) as usize,
-                                );
-                                *wave_out
-                                    .get_unchecked_mut(chan)
-                                    .as_mut()
-                                    .get_unchecked_mut(n) = interp_lin(frac_offset, buf);
-                            }
-                        }
-                    }
-                }
-            }
-            PolynomialDegree::Nearest => {
-                for n in 0..self.needed_output_size {
-                    t_ratio += t_ratio_increment;
-                    idx += t_ratio;
-                    let start_idx = idx.floor() as isize;
-                    for (chan, active) in self.channel_mask.iter().enumerate() {
-                        if *active {
-                            unsafe {
-                                let point = self.buffer.get_unchecked(chan).get_unchecked(
-                                    (start_idx + 2 * interpolator_len as isize) as usize,
-                                );
-                                *wave_out
-                                    .get_unchecked_mut(chan)
-                                    .as_mut()
-                                    .get_unchecked_mut(n) = *point;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Store last index for next iteration.
-        self.last_index = idx - self.needed_input_size as f64;
-        self.resample_ratio = self.target_ratio;
-        trace!(
-            "Resampling channels {:?}, {} frames in, {} frames out",
-            active_channels_mask,
-            self.needed_input_size,
-            self.needed_output_size,
-        );
-        let input_size = self.needed_input_size;
-        let output_size = self.needed_output_size;
-        self.update_lengths();
-        Ok((input_size, output_size))
-    }
-
-    fn output_frames_max(&self) -> usize {
-        Fast::<T>::calculate_max_output_size(
-            self.max_chunk_size,
-            self.resample_ratio_original,
-            self.max_relative_ratio,
-            &self.fixed,
-        )
-    }
-
-    fn output_frames_next(&self) -> usize {
-        self.needed_output_size
-    }
-
-    fn output_delay(&self) -> usize {
-        (POLYNOMIAL_LEN_U as f64 * self.resample_ratio / 2.0) as usize
-    }
-
-    fn nbr_channels(&self) -> usize {
-        self.nbr_channels
-    }
-
-    fn input_frames_max(&self) -> usize {
-        Fast::<T>::calculate_max_input_size(
-            self.max_chunk_size,
-            self.resample_ratio_original,
-            self.max_relative_ratio,
-            &self.fixed,
-        )
-    }
-
-    fn input_frames_next(&self) -> usize {
-        self.needed_input_size
-    }
-
-    fn set_resample_ratio(&mut self, new_ratio: f64, ramp: bool) -> ResampleResult<()> {
-        trace!("Change resample ratio to {}", new_ratio);
-        if (new_ratio / self.resample_ratio_original >= 1.0 / self.max_relative_ratio)
-            && (new_ratio / self.resample_ratio_original <= self.max_relative_ratio)
-        {
-            if !ramp {
-                self.resample_ratio = new_ratio;
-            }
-            self.target_ratio = new_ratio;
-            self.update_lengths();
-            Ok(())
-        } else {
-            Err(ResampleError::RatioOutOfBounds {
-                provided: new_ratio,
-                original: self.resample_ratio_original,
-                max_relative_ratio: self.max_relative_ratio,
-            })
-        }
-    }
-
-    fn set_resample_ratio_relative(&mut self, rel_ratio: f64, ramp: bool) -> ResampleResult<()> {
-        let new_ratio = self.resample_ratio_original * rel_ratio;
-        self.set_resample_ratio(new_ratio, ramp)
-    }
-
-    fn reset(&mut self) {
-        self.buffer
-            .iter_mut()
-            .for_each(|ch| ch.iter_mut().for_each(|s| *s = T::zero()));
-        self.channel_mask.iter_mut().for_each(|val| *val = true);
-        self.last_index = -(POLYNOMIAL_LEN_I / 2) as f64;
-        self.resample_ratio = self.resample_ratio_original;
-        self.target_ratio = self.resample_ratio_original;
-        self.chunk_size = self.max_chunk_size;
-        self.update_lengths();
-    }
-
-    fn set_chunk_size(&mut self, chunksize: usize) -> ResampleResult<()> {
-        if chunksize > self.max_chunk_size || chunksize == 0 {
-            return Err(ResampleError::InvalidChunkSize {
-                max: self.max_chunk_size,
-                requested: chunksize,
-            });
-        }
-        self.chunk_size = chunksize;
-        self.update_lengths();
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::PolynomialDegree;
-    use crate::Resampler;
-    use crate::{check_output, check_ratio};
-    use crate::{Fast, Fixed};
-    use rand::Rng;
-    use test_log::test;
-
-    #[test]
-    fn make_resampler_fi() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        let waves = vec![vec![0.0f64; 1024]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert!(
-            out[0].len() > 1150 && out[0].len() < 1229,
-            "expected {} - {} samples, got {}",
-            1150,
-            1229,
-            out[0].len()
-        );
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(out2.len(), 2, "Expected {} channels, got {}", 2, out2.len());
-        assert!(
-            out2[0].len() > 1226 && out2[0].len() < 1232,
-            "expected {} - {} samples, got {}",
-            1226,
-            1232,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn reset_resampler_fi() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-
-        let mut rng = rand::thread_rng();
-        let mut waves = vec![vec![0.0f64; 1024]; 2];
-        waves
-            .iter_mut()
-            .for_each(|ch| ch.iter_mut().for_each(|s| *s = rng.gen()));
-        let out1 = resampler.process(&waves, None).unwrap();
-        resampler.reset();
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(
-            out1, out2,
-            "Resampler gives different output when new and after a reset."
-        );
-    }
-
-    #[test]
-    fn make_resampler_fi_32() {
-        let mut resampler =
-            Fast::<f32>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        let waves = vec![vec![0.0f32; 1024]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert!(
-            out[0].len() > 1150 && out[0].len() < 1229,
-            "expected {} - {} samples, got {}",
-            1150,
-            1229,
-            out[0].len()
-        );
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(out2.len(), 2, "Expected {} channels, got {}", 2, out2.len());
-        assert!(
-            out2[0].len() > 1226 && out2[0].len() < 1232,
-            "expected {} - {} samples, got {}",
-            1226,
-            1232,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn make_resampler_fi_skipped() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        let waves = vec![vec![0.0f64; 1024], Vec::new()];
-        let mask = vec![true, false];
-        let out = resampler.process(&waves, Some(&mask)).unwrap();
-        assert_eq!(out.len(), 2);
-        assert!(out[0].len() > 1150 && out[0].len() < 1250);
-        assert!(out[1].is_empty());
-        let waves = vec![Vec::new(), vec![0.0f64; 1024]];
-        let mask = vec![false, true];
-        let out = resampler.process(&waves, Some(&mask)).unwrap();
-        assert_eq!(out.len(), 2);
-        assert!(out[1].len() > 1150 && out[0].len() < 1250);
-        assert!(out[0].is_empty());
-    }
-
-    #[test]
-    fn make_resampler_fi_downsample() {
-        // Replicate settings from reported issue.
-        let mut resampler = Fast::<f64>::new(
-            16000 as f64 / 96000 as f64,
-            1.0,
-            PolynomialDegree::Cubic,
-            1024,
-            2,
-            Fixed::Input,
-        )
-        .unwrap();
-        let waves = vec![vec![0.0f64; 1024]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert!(
-            out[0].len() > 140 && out[0].len() < 200,
-            "expected {} - {} samples, got {}",
-            140,
-            200,
-            out[0].len()
-        );
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(out2.len(), 2, "Expected {} channels, got {}", 2, out2.len());
-        assert!(
-            out2[0].len() > 167 && out2[0].len() < 173,
-            "expected {} - {} samples, got {}",
-            167,
-            173,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn make_resampler_fi_upsample() {
-        // Replicate settings from reported issue.
-        let mut resampler = Fast::<f64>::new(
-            192000 as f64 / 44100 as f64,
-            1.0,
-            PolynomialDegree::Cubic,
-            1024,
-            2,
-            Fixed::Input,
-        )
-        .unwrap();
-        let waves = vec![vec![0.0f64; 1024]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert!(
-            out[0].len() > 3800 && out[0].len() < 4458,
-            "expected {} - {} samples, got {}",
-            3800,
-            4458,
-            out[0].len()
-        );
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(out2.len(), 2, "Expected {} channels, got {}", 2, out2.len());
-        assert!(
-            out2[0].len() > 4455 && out2[0].len() < 4461,
-            "expected {} - {} samples, got {}",
-            4455,
-            4461,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn make_resampler_fo() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-        println!("{}", frames);
-        assert!(frames > 800 && frames < 900);
-        let waves = vec![vec![0.0f64; frames]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[0].len(), 1024);
-    }
-
-    #[test]
-    fn reset_resampler_fo() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-
-        let mut rng = rand::thread_rng();
-        let mut waves = vec![vec![0.0f64; frames]; 2];
-        waves
-            .iter_mut()
-            .for_each(|ch| ch.iter_mut().for_each(|s| *s = rng.gen()));
-        let out1 = resampler.process(&waves, None).unwrap();
-        resampler.reset();
-        assert_eq!(
-            frames,
-            resampler.input_frames_next(),
-            "Resampler requires different number of frames when new and after a reset."
-        );
-        let out2 = resampler.process(&waves, None).unwrap();
-        assert_eq!(
-            out1, out2,
-            "Resampler gives different output when new and after a reset."
-        );
-    }
-
-    #[test]
-    fn make_resampler_fo_32() {
-        let mut resampler =
-            Fast::<f32>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-        println!("{}", frames);
-        assert!(frames > 800 && frames < 900);
-        let waves = vec![vec![0.0f32; frames]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[0].len(), 1024);
-    }
-
-    #[test]
-    fn make_resampler_fo_skipped() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-        println!("{}", frames);
-        assert!(frames > 800 && frames < 900);
-        let mut waves = vec![vec![0.0f64; frames], Vec::new()];
-        let mask = vec![true, false];
-        waves[0][100] = 3.0;
-        let out = resampler.process(&waves, Some(&mask)).unwrap();
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[0].len(), 1024);
-        assert!(out[1].is_empty());
-        println!("{:?}", out[0]);
-        let summed = out[0].iter().sum::<f64>();
-        println!("sum: {}", summed);
-        assert!(summed < 4.0);
-        assert!(summed > 2.0);
-
-        let frames = resampler.input_frames_next();
-        let mut waves = vec![Vec::new(), vec![0.0f64; frames]];
-        let mask = vec![false, true];
-        waves[1][10] = 3.0;
-        let out = resampler.process(&waves, Some(&mask)).unwrap();
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[1].len(), 1024);
-        assert!(out[0].is_empty());
-        let summed = out[1].iter().sum::<f64>();
-        assert!(summed < 4.0);
-        assert!(summed > 2.0);
-    }
-
-    #[test]
-    fn make_resampler_fo_downsample() {
-        let mut resampler =
-            Fast::<f64>::new(0.125, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-        println!("{}", frames);
-        assert!(
-            frames > 8192 && frames < 9000,
-            "expected {}..{} samples, got {}",
-            8192,
-            9000,
-            frames
-        );
-        let waves = vec![vec![0.0f64; frames]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert_eq!(
-            out[0].len(),
-            1024,
-            "Expected {} frames, got {}",
-            1024,
-            out[0].len()
-        );
-        let frames2 = resampler.input_frames_next();
-        assert!(
-            frames2 > 8189 && frames2 < 8195,
-            "expected {}..{} samples, got {}",
-            8189,
-            8195,
-            frames2
-        );
-        let waves2 = vec![vec![0.0f64; frames2]; 2];
-        let out2 = resampler.process(&waves2, None).unwrap();
-        assert_eq!(
-            out2[0].len(),
-            1024,
-            "Expected {} frames, got {}",
-            1024,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn make_resampler_fo_upsample() {
-        let mut resampler =
-            Fast::<f64>::new(8.0, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        let frames = resampler.input_frames_next();
-        println!("{}", frames);
-        assert!(
-            frames > 128 && frames < 300,
-            "expected {}..{} samples, got {}",
-            140,
-            200,
-            frames
-        );
-        let waves = vec![vec![0.0f64; frames]; 2];
-        let out = resampler.process(&waves, None).unwrap();
-        assert_eq!(out.len(), 2, "Expected {} channels, got {}", 2, out.len());
-        assert_eq!(
-            out[0].len(),
-            1024,
-            "Expected {} frames, got {}",
-            1024,
-            out[0].len()
-        );
-        let frames2 = resampler.input_frames_next();
-        assert!(
-            frames2 > 125 && frames2 < 131,
-            "expected {}..{} samples, got {}",
-            125,
-            131,
-            frames2
-        );
-        let waves2 = vec![vec![0.0f64; frames2]; 2];
-        let out2 = resampler.process(&waves2, None).unwrap();
-        assert_eq!(
-            out2[0].len(),
-            1024,
-            "Expected {} frames, got {}",
-            1024,
-            out2[0].len()
-        );
-    }
-
-    #[test]
-    fn check_fo_output_up() {
-        let mut resampler =
-            Fast::<f64>::new(8.0, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        check_output!(resampler);
-    }
-
-    #[test]
-    fn check_fo_output_down() {
-        let mut resampler =
-            Fast::<f64>::new(0.8, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        check_output!(resampler);
-    }
-
-    #[test]
-    fn check_fi_output_up() {
-        let mut resampler =
-            Fast::<f64>::new(8.0, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        check_output!(resampler);
-    }
-
-    #[test]
-    fn check_fi_output_down() {
-        let mut resampler =
-            Fast::<f64>::new(0.8, 1.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        check_output!(resampler);
-    }
-
-    #[test]
-    fn resample_small_fo_up() {
-        let ratio = 96000.0 / 44100.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1, 2, Fixed::Output).unwrap();
-        check_ratio!(resampler, ratio, 1000000);
-    }
-
-    #[test]
-    fn resample_big_fo_up() {
-        let ratio = 96000.0 / 44100.0;
-        let mut resampler = Fast::<f32>::new(
-            ratio,
-            100.0,
-            PolynomialDegree::Cubic,
-            1024,
-            2,
-            Fixed::Output,
-        )
-        .unwrap();
-        check_ratio!(resampler, ratio, 1000);
-    }
-
-    #[test]
-    fn resample_small_fo_down() {
-        let ratio = 44100.0 / 96000.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1, 2, Fixed::Output).unwrap();
-        check_ratio!(resampler, ratio, 1000000);
-    }
-
-    #[test]
-    fn resample_big_fo_down() {
-        let ratio = 44100.0 / 96000.0;
-        let mut resampler = Fast::<f32>::new(
-            ratio,
-            100.0,
-            PolynomialDegree::Cubic,
-            1024,
-            2,
-            Fixed::Output,
-        )
-        .unwrap();
-        check_ratio!(resampler, ratio, 1000);
-    }
-
-    #[test]
-    fn resample_small_fi_up() {
-        let ratio = 96000.0 / 44100.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1, 2, Fixed::Input).unwrap();
-        check_ratio!(resampler, ratio, 1000000);
-    }
-
-    #[test]
-    fn resample_big_fi_up() {
-        let ratio = 96000.0 / 44100.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        check_ratio!(resampler, ratio, 1000);
-    }
-
-    #[test]
-    fn resample_small_fi_down() {
-        let ratio = 44100.0 / 96000.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1, 2, Fixed::Input).unwrap();
-        check_ratio!(resampler, ratio, 1000000);
-    }
-
-    #[test]
-    fn resample_big_fi_down() {
-        let ratio = 44100.0 / 96000.0;
-        let mut resampler =
-            Fast::<f32>::new(ratio, 100.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        check_ratio!(resampler, ratio, 1000);
-    }
-
-    #[test]
-    fn check_fo_output_resize() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 100.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Output).unwrap();
-        assert_eq!(resampler.output_frames_next(), 1024);
-        resampler.set_chunk_size(256).unwrap();
-        assert_eq!(resampler.output_frames_next(), 256);
-        check_output!(resampler);
-    }
-
-    #[test]
-    fn check_fi_output_resize() {
-        let mut resampler =
-            Fast::<f64>::new(1.2, 100.0, PolynomialDegree::Cubic, 1024, 2, Fixed::Input).unwrap();
-        assert_eq!(resampler.input_frames_next(), 1024);
-        resampler.set_chunk_size(256).unwrap();
-        assert_eq!(resampler.input_frames_next(), 256);
-        check_output!(resampler);
-    }
 }
